requirements.txt
# Core framework
google-adk>=1.18.0
litellm>=1.50.0
groq>=0.9.0

# Environment management
python-dotenv>=1.0.0

# Web search
googlesearch-python>=1.3.0

# Server
uvicorn>=0.30.0
fastapi>=0.115.0
